A more general MAP/MLE optimisation interface
Added functions to create optimisation problems from a Turing model both in constrained and unconstrained space which can be passed to any Julia optimisation library such as `BlackBoxOptim.jl`. Also added a `GalacticOptim`-specific function which creates a `GalacticOptim.OptimisationFunction`. Added a series  of test for both implementation and the interaction with `GalacticOptim`.