Minor refactor of MLE/MAP code
Added DiffEqBase in order to create OptimisationFunctions without the need of loading GalacticOptim or having it as dependency. Moved `Optim.jl` specific code into `mode/OptimInterface.jl` which is only loaded once`Optim.jl` is being used keeping backwards compatibility and keeping the calculations of summary stats of the parameter estimates. Moved all the "generic" optimisation code into `mode/ModeEstimation.jl`. MLE() and MAP() create objectives in the constrained space and MLE(false) and MAP(false) create objectives in the unconstrained space.

All test pass but I am not 100% confident on all the import and using statements to be used in the most elegant and coherent way.