Better describe the restrictions when using reversediff
When compiling the tape with the ReverseDiff system you can't use certain kinds of models. Previously the documentation had just said no to runtime if statements. However this is a HUGE restriction and not actually precise. The key is that at each step of sampling, the same instructions need to be executed each time, thus the same branches need to get executed. This will occur when the if statement can be determined at compile time, but will also occur whenever the if statements depend only on the data passed to the model, as the data is constant throughout the sampling. This is quite significant, because in the absence of the ability to use if statements based on the data, you can't even express models where you treat say males differently from females using an if to test for sex. (or any number of similar situations, such as treating different measurement instruments with different measurement error, or the like). 

By making the documentation more precise users will have less trouble determining what kinds of models they can use and will be able to safely use ReverseDiff on a much expanded set of models.

See discussion in this thread: https://discourse.julialang.org/t/clarification-on-setrdcache-in-turing/88767/6?u=dlakelan