Fix error TuringOptimExt.jl with information matrix/vcov (#2167)
* Update TuringOptimExt.jl

Fix errors in information matrix/variance-covariance matrix calculation for MLEs/MAPs

* Update OptimInterface.jl: use vcov instead of informationmatrix in linear regression test

* Update OptimInterface.jl: fix StatsBase integration tests

* Update TuringOptimExt.jl, fix calculation of info matrix

Usually, the information matrix is the negative of the Hessian of the log likelihood, but because Optim minimizes the negative of the log likelihood, we don't need to take the negative here. The Hessian is the info matrix in this context.

* Update OptimInterface.jl: calculate explicit info and vcov matrices in test

The analytical values for the information and vcov matrices are now calculated using the standard equations. Added separate tests for both matrices.

* Fix silly error in calculation of info/vcov matrix. Correct linear regression test.

---------

Co-authored-by: Hong Ge <3279477+yebai@users.noreply.github.com>