make MCMCThreads etc. respect `check_model=false` (#2721)
Prior to this PR, if you called

```julia
sample(model, spl, MCMCThreads(), N, n; check_model=false)
```

the model checking here


https://github.com/TuringLang/Turing.jl/blob/ed7f76c7221a756a390d3168ed6e1fcb6f95d263/src/mcmc/abstractmcmc.jl#L121

would be correctly skipped. The problem is that this calls
`AbstractMCMC.mcmcsample`, which then calls the single-threaded
`sample`, but the `check_model` argument is not passed down, so it
defaults to `true` which then checks the model(!!)

This PR fixes it by simply passing `check_model=false` through. (The
point of explicitly setting it to false is that if the user wanted the
check, it would be done on the line above already, there's no need to
check the model `N+1` times.)

The added test fails on main and passes on this PR.