Update for DynamicPPL 0.39 (#2715)
The main change in DPPL 0.39 is OnlyAccsVarInfo and faster evaluation.

This PR uses fast evaluation in MCMC sampling where it can. MCMC
sampling mostly works as can be seen from the tests.

## Of note

- MCMCChains no longer stores an overall `chain.logevidence` field

The reason is because we now use [the `bundle_samples` method in
DynamicPPL](https://github.com/TuringLang/DynamicPPL.jl/blob/a6d56a2b9074d9da27eea4a6e4a2ab9a3013913f/ext/DynamicPPLMCMCChainsExt.jl#L121-L156),
which has no way of reliably determining the log-evidence from the
transition. If we wanted to fix this, we would have to add a
`getlogevidence` function in AbstractMCMC.

I personally don't consider this a problem. The reason why log-evidence
used to be stored was because chains did not provide enough information
for people to calculate this themselves (specifically, chains only
stored logp, not loglikelihood). Now that `chn[:likelihood]` contains
the likelihood, it should be ok for people to calculate this themselves.

---------

Co-authored-by: Markus Hauru <markus@mhauru.org>