Use OnlyAccsVarInfo in check_model (#2742)
Because the model is contextualised with InitContext, the call to
check_model doesn't actually need anything more than an OnlyAccsVarInfo.
This allows models that don't work well with Metadata* to still be
checked for validity (note that the check, as defined in DynamicPPL,
only uses DebugAccumulator, it does not actually use metadata in any
form -- unless the model has DefaultContext as its leaf context -- which
is not the case here).

(cc @PTWaade I think this is the reason why check_model failed)

\* one might ask what models don't work well with metadata, but still
otherwise work. There are a couple of things that come to mind: 1)
parallel assume statements; and 2) special distribution types for which
the from_[linked_]_vec_transform functions are not defined. These models
should in principle work with `Prior()` sampling, which only ever uses
OnlyAccsVarInfo. However, currently, they don't unless you disable
check_model; this is the behaviour that this PR fixes.